requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
ecdsa==0.18.0
filelock==3.10.7
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
platformdirs==3.2.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.21.0
